remove SwitchBackendConfirmationComponent
