+++ WireAuthentication/Sources/WireAuthentication/Needle.generated.swift
@@ -109,37 +109,6 @@ private class DetermineAuthMethodComponentDependency527e70b5dbcfcb8f2023Provider
 private func factoryd47fa74281e135cd9f10b3a8f24c1d289f2c0f2e(_ component: NeedleFoundation.Scope) -> AnyObject {
     return DetermineAuthMethodComponentDependency527e70b5dbcfcb8f2023Provider(rootComponent: parent1(component) as! RootComponent)
 }
-private class SwitchBackendConfirmationComponentDependency7a1956d88810c08ef169Provider: SwitchBackendConfirmationComponentDependency {
-    var router: any Router {
-        return rootComponent.router
-    }
-    var preferredAPIVersion: APIVersion? {
-        return rootComponent.preferredAPIVersion
-    }
-    var productionVersions: Set<APIVersion> {
-        return rootComponent.productionVersions
-    }
-    var minTLSVersion: TLSVersion {
-        return rootComponent.minTLSVersion
-    }
-    var ssoCallbackURLScheme: String {
-        return rootComponent.ssoCallbackURLScheme
-    }
-    var userDefaults: UserDefaults {
-        return rootComponent.userDefaults
-    }
-    var appStoreURL: URL {
-        return rootComponent.appStoreURL
-    }
-    private let rootComponent: RootComponent
-    init(rootComponent: RootComponent) {
-        self.rootComponent = rootComponent
-    }
-}
-/// ^->RootComponent->DetermineAuthMethodComponent->SwitchBackendConfirmationComponent
-private func factorye1144df20d596f07c3bea9403e3301bb54f80df0(_ component: NeedleFoundation.Scope) -> AnyObject {
-    return SwitchBackendConfirmationComponentDependency7a1956d88810c08ef169Provider(rootComponent: parent2(component) as! RootComponent)
-}
 private class LoginViaSSODependencycb22423a897409b8b5faProvider: LoginViaSSODependency {
 
 
@@ -177,10 +146,6 @@ private func factory5f94de319ad3e04a942342f5655bf2362a8495f6(_ component: Needle
 /// ^->RootComponent->DetermineAuthMethodComponent->NoHistoryComponent
 private func factory5f94de319ad3e04a9423a9403e3301bb54f80df0(_ component: NeedleFoundation.Scope) -> AnyObject {
     return NoHistoryComponentDependencya1005f718577ea03ea08Provider(rootComponent: parent2(component) as! RootComponent)
-}
-/// ^->RootComponent->NoHistoryComponent
-private func factory5f94de319ad3e04a9423b3a8f24c1d289f2c0f2e(_ component: NeedleFoundation.Scope) -> AnyObject {
-    return NoHistoryComponentDependencya1005f718577ea03ea08Provider(rootComponent: parent1(component) as! RootComponent)
 }
 private class DetermineAuthMethodOnPremComponentDependencyb4a02ddb77e5a66140fbProvider: DetermineAuthMethodOnPremComponentDependency {
     var router: any Router {
@@ -280,18 +245,6 @@ extension DetermineAuthMethodComponent: NeedleFoundation.Registration {
         localTable["networkService-NetworkService"] = { [unowned self] in self.networkService as Any }
     }
 }
-extension SwitchBackendConfirmationComponent: NeedleFoundation.Registration {
-    public func registerItems() {
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.router] = "router-any Router"
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.preferredAPIVersion] = "preferredAPIVersion-APIVersion?"
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.productionVersions] = "productionVersions-Set<APIVersion>"
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.minTLSVersion] = "minTLSVersion-TLSVersion"
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.ssoCallbackURLScheme] = "ssoCallbackURLScheme-String"
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.userDefaults] = "userDefaults-UserDefaults"
-        keyPathToName[\SwitchBackendConfirmationComponentDependency.appStoreURL] = "appStoreURL-URL"
-        localTable["backendConfig-BackendConfig"] = { [unowned self] in self.backendConfig as Any }
-    }
-}
 extension LoginViaSSOComponent: NeedleFoundation.Registration {
     public func registerItems() {
 
@@ -373,19 +326,17 @@ private func registerProviderFactory(_ componentPath: String, _ factory: @escapi
     registerProviderFactory("^->RootComponent->LoginViaEmailComponent->VerificationCodeComponent", factoryd3638676a47fce1fe623a054ff50c10da855de8a)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->LoginViaEmailComponent->VerificationCodeComponent", factoryd3638676a47fce1fe62317031e1ba787d83cb463)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent", factoryd47fa74281e135cd9f10b3a8f24c1d289f2c0f2e)
-    registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->SwitchBackendConfirmationComponent", factorye1144df20d596f07c3bea9403e3301bb54f80df0)
-    registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->SwitchBackendConfirmationComponent", factorye1144df20d596f07c3bea9403e3301bb54f80df0)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->LoginViaSSOComponent", factory075263b25e612b6948d3e3b0c44298fc1c149afb)
-    registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->SwitchBackendConfirmationComponent->LoginViaSSOComponent", factory075263b25e612b6948d3e3b0c44298fc1c149afb)
-    registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->SwitchBackendConfirmationComponent->LoginViaSSOComponent", factory075263b25e612b6948d3e3b0c44298fc1c149afb)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->LoginViaSSOComponent", factory075263b25e612b6948d3e3b0c44298fc1c149afb)
     registerProviderFactory("^->RootComponent", factoryEmptyDependencyProvider)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->LoginViaEmailComponent->VerificationCodeComponent->NoHistoryComponent", factory5f94de319ad3e04a942321a9c45ed079aafca21f)
     registerProviderFactory("^->RootComponent->LoginViaEmailComponent->VerificationCodeComponent->NoHistoryComponent", factory5f94de319ad3e04a942342f5655bf2362a8495f6)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->LoginViaEmailComponent->VerificationCodeComponent->NoHistoryComponent", factory5f94de319ad3e04a942321a9c45ed079aafca21f)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->NoHistoryComponent", factory5f94de319ad3e04a9423a9403e3301bb54f80df0)
-    registerProviderFactory("^->RootComponent->NoHistoryComponent", factory5f94de319ad3e04a9423b3a8f24c1d289f2c0f2e)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->NoHistoryComponent", factory5f94de319ad3e04a9423a9403e3301bb54f80df0)
+    registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->LoginViaEmailComponent->NoHistoryComponent", factory5f94de319ad3e04a942342f5655bf2362a8495f6)
+    registerProviderFactory("^->RootComponent->LoginViaEmailComponent->NoHistoryComponent", factory5f94de319ad3e04a9423a9403e3301bb54f80df0)
+    registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent->LoginViaEmailComponent->NoHistoryComponent", factory5f94de319ad3e04a942342f5655bf2362a8495f6)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodOnPremComponent", factorydbdff85f3341dce5e925b3a8f24c1d289f2c0f2e)
     registerProviderFactory("^->RootComponent->DetermineAuthMethodComponent->LoginViaEmailComponent", factory9bda312c16141c932061a9403e3301bb54f80df0)
     registerProviderFactory("^->RootComponent->LoginViaEmailComponent", factory9bda312c16141c932061b3a8f24c1d289f2c0f2e)
