--- WireAuthentication/Sources/WireAuthenticationUI/Views/SwitchBackend/SwitchBackendConfirmationViewModel.swift
@@ -1,209 +1,0 @@
-//
-// Wire
-// Copyright (C) 2025 Wire Swiss GmbH
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-//
-
-import Foundation
-import WireAuthenticationAPI
-import WireLogging
-
-@MainActor
-package class SwitchBackendConfirmationViewModel: ObservableObject {
-
-    package enum ModalDestination: Hashable, Identifiable, Sendable {
-        package var id: Self { self }
-
-        case ssoLogin(url: URL, backendEnvironment: WireAuthenticationBackendEnvironment)
-    }
-
-    private typealias Strings = L10n.SwitchBackendConfirmation
-
-    package typealias Factory =
-        FetchSSOURLUseCaseFactory &
-        OpenAppStoreUseCaseFactory &
-        ResolveBackendMetadataUseCaseFactory
-
-    // MARK: - State
-
-    let items: [ItemUIModel]
-
-    private let router: any Router
-    private let factory: any Factory
-    private let email: String?
-    private let environmentType: BackendEnvironmentType
-    private let backendConfig: BackendConfig
-
-    @Published private(set) var isLoading = false
-    @Published var alert: Alert?
-    @Published var modalDestination: ModalDestination?
-
-    // MARK: - Life cycle
-
-    package init(
-        router: any Router,
-        factory: any Factory,
-        email: String?,
-        environmentType: BackendEnvironmentType,
-        backendConfig: BackendConfig
-    ) {
-        self.router = router
-        self.factory = factory
-        self.email = email
-        self.environmentType = environmentType
-        self.backendConfig = backendConfig
-
-        var items = [
-            ItemUIModel(
-                title: Strings.backendName,
-                value: backendConfig.title,
-                isURL: false
-            ),
-            ItemUIModel(
-                title: Strings.backendUrl,
-                value: backendConfig.endpoints.backendURL.absoluteString,
-                isURL: true
-            ),
-            ItemUIModel(
-                title: Strings.backendWsurl,
-                value: backendConfig.endpoints.backendWSURL.absoluteString,
-                isURL: true
-            ),
-            ItemUIModel(
-                title: Strings.blacklistUrl,
-                value: backendConfig.endpoints.blackListURL.absoluteString,
-                isURL: true
-            ),
-            ItemUIModel(
-                title: Strings.teamsUrl,
-                value: backendConfig.endpoints.teamsURL.absoluteString,
-                isURL: true
-            ),
-            ItemUIModel(
-                title: Strings.accountsUrl,
-                value: backendConfig.endpoints.accountsURL.absoluteString,
-                isURL: true
-            ),
-            ItemUIModel(
-                title: Strings.websiteUrl,
-                value: backendConfig.endpoints.websiteURL.absoluteString,
-                isURL: true
-            )
-        ]
-
-        if let countlyURL = backendConfig.endpoints.countlyURL {
-            items.append(
-                ItemUIModel(
-                    title: Strings.countlyUrl,
-                    value: countlyURL.absoluteString,
-                    isURL: true
-                )
-            )
-        }
-
-        self.items = items
-    }
-
-    func confirm() async {
-        isLoading = true
-
-        defer {
-            isLoading = false
-        }
-
-        // If authenticated proxy is required, go straight to email login because we need to
-        // get proxy credentials first.
-        if let proxySettings = backendConfig.proxySettings, proxySettings.needsAuthentication {
-//            router.presentSheet(
-//                RootView.ModalDestination.onPremiseLogin(
-//                    email: email,
-//                    environmentType: environmentType,
-//                    environment: backendConfig,
-//                    backendMetadata: nil
-//                )
-//            )
-        } else {
-            do {
-                // Before we can make requests we need to resolve the api version.
-                let backendMetadata = try await resolveBackendMetadata()
-
-                if let ssoURL = try await fetchSSOURL(apiVersion: backendMetadata.apiVersion) {
-                    let backendEnvironment = WireAuthenticationBackendEnvironment(
-                        environmentType: environmentType,
-                        config: backendConfig,
-                        metadata: backendMetadata
-                    )
-
-                    modalDestination = .ssoLogin(url: ssoURL, backendEnvironment: backendEnvironment)
-                    WireLogger.authentication.info("Fetching default SSO URL succeed")
-                } else if let email {
-                    router.presentSheet(
-                        RootView.ModalDestination.onPremiseLogin(
-                            email: email,
-                            environmentType: environmentType,
-                            environment: backendConfig,
-                            backendMetadata: backendMetadata
-                        )
-                    )
-                } else {
-                    router.presentSheet(
-                        RootView.ModalDestination.onPremiseAuthFlow(
-                            environmentType: environmentType,
-                            backendConfig: backendConfig,
-                            backendMetadata: backendMetadata
-                        )
-                    )
-                }
-            } catch ResolveBackendMetadataUseCaseFailure.clientVersionObsolete {
-                WireLogger.authentication.error("detected obsolete client")
-                alert = .obsoleteClient
-            } catch ResolveBackendMetadataUseCaseFailure.backendAPIVersionObsolete {
-                WireLogger.authentication.error("detected obsolete backend")
-                alert = .obsoleteBackend
-            } catch {
-                WireLogger.authentication.error("Fetching default SSO URL failed: \(error)")
-                alert = .general(for: error)
-            }
-        }
-    }
-
-    private func resolveBackendMetadata() async throws -> BackendMetadata {
-        let useCase = factory.resolveBackendMetadataUseCase()
-        return try await Task.detached {
-            try await useCase.invoke()
-        }.value
-    }
-
-    private func fetchSSOURL(apiVersion: WireAuthenticationAPI.BackendMetadata.APIVersion) async throws -> URL? {
-        let useCase = factory.fetchSSOURLUseCase(apiVersion: apiVersion)
-        return try await Task.detached {
-            try await useCase.invoke()
-        }.value
-    }
-
-    func goToAppStore() {
-        factory.openAppStoreUseCase().invoke()
-        alert = nil
-    }
-
-    // MARK: - Model
-
-    package struct ItemUIModel {
-        let title: String
-        let value: String
-        let isURL: Bool
-    }
-
-}
